pyproject.toml
[project]
name = "hk-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
